requirements.txt
pandas
yfinance
click
python-dotenv
pytest
pyarrow
streamlit
ccxt
openai
